Extract introspect out of expo-cli
